victoria_backend/requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
astroid==2.2.5
autopep8==1.4.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.1
Django==3.0.3
django-rest-framework==0.1.0
djangorestframework==3.11.0
filelock==3.0.12
idna==2.10
isort==4.3.17
joblib==0.16.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.19.1
packaging==20.4
pycodestyle==2.5.0
pylint==2.3.1
pyparsing==2.4.7
pytz==2020.1
regex==2020.7.14
requests==2.24.0
sacremoses==0.0.43
sentencepiece==0.1.91
six==1.15.0
sqlparse==0.3.1
tokenizers==0.8.1rc1
tqdm==4.48.0
transformers==3.0.2
typed-ast==1.3.4
urllib3==1.25.10
wincertstore==0.2
wrapt==1.11.1
